+++ feature/poster/src/main/kotlin/com/example/cinema/feature/poster/data/network/PosterApi.kt
@@ -1,11 +1,9 @@
 package com.example.cinema.feature.poster.data.network
 
-import com.example.cinema.feature.poster.data.model.FilmModel
 import com.example.cinema.feature.poster.data.model.PosterModel
 import retrofit2.http.GET
-import retrofit2.http.Path
 
-interface FilmApi {
+interface PosterApi {
 
 	@GET("today")
 	suspend fun getAll(): PosterModel
