+++ feature/poster/src/main/kotlin/com/example/cinema/feature/poster/domain/repository/PosterRepository.kt
@@ -2,7 +2,7 @@ package com.example.cinema.feature.poster.domain.repository
 
 import com.example.cinema.feature.poster.domain.entity.Film
 
-interface FilmRepository {
+interface PosterRepository {
 
 	suspend fun getAllFilmsRepository(): List<Film>
 }
